requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.11.0
attrs==25.4.0
awscli==1.41.14
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.39.14
botocore==1.39.14
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
distlib==0.4.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docker==7.1.0
docutils==0.19
ecdsa==0.19.1
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.1
fsspec==2025.7.0
gay==1.3.4
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==8.7
keyrings.alt==5.0.2
kiwisolver==1.4.9
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.5
more-itertools==10.7.0
mpmath==1.3.0
narwhals==2.7.0
networkx==3.5
numpy==2.2.6
openai==1.97.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.5.0
protobuf==6.32.1
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.1
rsa==4.7.2
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.4
starlette==0.47.3
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.184
ultralytics-thop==2.0.16
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.35.3
voluptuous==0.15.2
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.11.0
yolo==0.3.1
